+++ src/pipeline_telemetry/__init__.py
@@ -41,7 +41,8 @@ from errors import ListErrors
 from .decorator import add_mongo_single_usage_telemetry, add_mongo_telemetry, \
     add_single_usage_telemetry, add_telemetry
 from .helper import add_errors_from_return_value, increase_base_count, \
-    increase_fail_count, is_telemetry_counter
+    increase_fail_count, is_telemetry_counter, process_return_value, \
+    process_telemetry_counters_in_return_value
 from .main import Telemetry
 from .settings.data_class import ProcessType, TelemetryCounter
 from .settings.process_type import ProcessTypes